[#658] Centralize helper methods from view to logic
* Removed redundant helper methods from view_helpers.rb
* Introduced logic_helpers.rb for shared logic
* Ensured consistent utility in both views and logic modules
