+++ lib/onetime/utils.rb
@@ -59,4 +59,96 @@ module Onetime
       text.gsub regex, '\\3*****\\4@\\6*****\\7'
     end
   end
+
+  module TimeUtils
+
+      def epochdate(time_in_s)
+        time_parsed = Time.at time_in_s.to_i
+        dformat time_parsed.utc
+      end
+
+      def epochtime(time_in_s)
+        time_parsed = Time.at time_in_s.to_i
+        tformat time_parsed.utc
+      end
+
+      def epochformat(time_in_s)
+        time_parsed = Time.at time_in_s.to_i
+        dtformat time_parsed.utc
+      end
+
+      def epochformat2(time_in_s)
+        time_parsed = Time.at time_in_s.to_i
+        dtformat2 time_parsed.utc
+      end
+
+      def epochdom(time_in_s)
+        time_parsed = Time.at time_in_s.to_i
+        time_parsed.utc.strftime('%b %d, %Y')
+      end
+
+      def epochtod(time_in_s)
+        time_parsed = Time.at time_in_s.to_i
+        time_parsed.utc.strftime('%I:%M%p').gsub(/^0/, '').downcase
+      end
+
+      def epochcsvformat(time_in_s)
+        time_parsed = Time.at time_in_s.to_i
+        time_parsed.utc.strftime('%Y/%m/%d %H:%M:%S')
+      end
+
+      def dtformat(time_in_s)
+        time_in_s = DateTime.parse time_in_s unless time_in_s.is_a?(Time)
+        time_in_s.strftime('%Y-%m-%d@%H:%M:%S UTC')
+      end
+
+      def dtformat2(time_in_s)
+        time_in_s = DateTime.parse time_in_s unless time_in_s.is_a?(Time)
+        time_in_s.strftime('%Y-%m-%d@%H:%M UTC')
+      end
+
+      def dformat(time_in_s)
+        time_in_s = DateTime.parse time_in_s unless time_in_s.is_a?(Time)
+        time_in_s.strftime('%Y-%m-%d')
+      end
+
+      def tformat(time_in_s)
+        time_in_s = DateTime.parse time_in_s unless time_in_s.is_a?(Time)
+        time_in_s.strftime('%H:%M:%S')
+      end
+
+      # rubocop:disable Metrics/PerceivedComplexity, Metrics/AbcSize
+      def natural_time(time_in_s)
+        return if time_in_s.nil?
+
+        val = Time.now.utc.to_i - time_in_s.to_i
+        # puts val
+        if val < 10
+          result = 'a moment ago'
+        elsif val < 40
+          result = "about #{(val * 1.5).to_i.to_s.slice(0, 1)}0 seconds ago"
+        elsif val < 60
+          result = 'about a minute ago'
+        elsif val < 60 * 1.3
+          result = '1 minute ago'
+        elsif val < 60 * 2
+          result = '2 minutes ago'
+        elsif val < 60 * 50
+          result = "#{(val / 60).to_i} minutes ago"
+        elsif val < 3600 * 1.4
+          result = 'about 1 hour ago'
+        elsif val < 3600 * (24 / 1.02)
+          result = "about #{(val / 60 / 60 * 1.02).to_i} hours ago"
+        elsif val < 3600 * 24 * 1.6
+          result = Time.at(time_in_s.to_i).strftime('yesterday').downcase
+        elsif val < 3600 * 24 * 7
+          result = Time.at(time_in_s.to_i).strftime('on %A').downcase
+        else
+          weeks = (val / 3600.0 / 24.0 / 7).to_i
+          result = Time.at(time_in_s.to_i).strftime("#{weeks} #{'week'.plural(weeks)} ago").downcase
+        end
+        result
+      end
+
+  end
 end
